requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
gunicorn==20.1.0
Flask-SQLAlchemy==2.5.1
PyMySQL==0.9.3
google-api-core==2.10.2
google-api-python-client==1.5.1
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==1.7.3
google-cloud-dns==0.34.1
google-cloud-resource-manager==0.30.0
googleapis-common-protos==1.57.0
oauth2client==2.2.0
SQLAlchemy==1.4.44
